requirements.txt
streamlit==1.45.1
pandas==2.2.3
pydantic-ai==0.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
appnope==0.1.4
argcomplete==3.6.2
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.13
botocore==1.38.13
cachetools==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.24.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
identify==2.6.10
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.3.42
logfire==3.15.1
logfire-api==3.15.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.8.0
mdurl==0.1.2
mistralai==1.7.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.38.2
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.2.5
openai==1.78.0
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-http==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-ai==0.1.11
pydantic-ai-slim==0.1.11
pydantic-evals==0.1.11
pydantic-graph==0.1.11
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0